% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_hpo_id_column_to_rd_ewce_results.R
\name{add_hpo_termid_col}
\alias{add_hpo_termid_col}
\title{Add HPO term Id column to dataframe.}
\usage{
add_hpo_termid_col(cells, phenotype_to_genes, hpo)
}
\arguments{
\item{cells}{The dataframe of subset of RD EWCE results to be plotted in the cell select app.}

\item{phenotype_to_genes}{The hpo terms with gene list annotations data frame from hpo website}

\item{hpo}{The HPO ontology data object}
}
\value{
The subset of ewce result data frame with a HPO Id column added.
}
\description{
This adds the HPO term id column to the subest of ewce results data to be plotted
in the cell select app. It also checks if it is a valid HPO term id to pevent error and adds
a boolean column where TRUE if term is valid. If the HPO Id is not correct, it caused
an error in the ontologyPlot package
}
\examples{
\dontrun{
add_hpo_termid_col(cells,phenotype_to_genes,hpo)}
}
