% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_enrichments_per_cell_type_dataframe.R
\name{dataframe_phenos_per_cell}
\alias{dataframe_phenos_per_cell}
\title{Dataframe of number of significant RD EWCE Phenotypes per cell type}
\usage{
dataframe_phenos_per_cell(
  all_results_merged,
  cell_mappings,
  fold = 1,
  q_val = 0.005,
  dataset = "Descartes"
)
}
\arguments{
\item{all_results_merged}{Data frame of RD EWCE results}

\item{cell_mappings}{tissue-cell mappings for colouring results by tissue type in the plot}

\item{fold}{Fold threshold for deciding which results count as significant}

\item{q_val}{The q value threshold for significant results}

\item{dataset}{"Descartes" or "other" string. Probably not needed anymore}
}
\value{
A data frame of enrichment counts per cell
}
\description{
This creates the dataframe of n signif enrichments for each cell type from the
RD EWCE results. It is used in the plot_phenos_per_cell function. It also orders the
results by number of phenotypes for better plot.
}
