% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ewce_plot_function.R
\name{ewce.plot}
\alias{ewce.plot}
\title{Modification to ewce.plot function to automatically align the dendrogram}
\usage{
ewce.plot(total_res, mtc_method = "bonferroni", ctd = NULL)
}
\arguments{
\item{total_res}{Results to be plotted ?}

\item{mtc_method}{Multiple comparison adjustment method}

\item{ctd}{Cell type data object <list>}
}
\value{
A bar chart with dendrogram of EWCE results in each cell type.
}
\description{
Modification to ewce.plot function to automatically align the dendrogram
}
\examples{
\dontrun{ewce.plot(results_dataframe,"bonferroni",CTD)}
}
