% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_get_relative_ont_level.R
\name{find_parent}
\alias{find_parent}
\title{Identify relative ontology level for HPO term in a subset of ontology}
\usage{
find_parent(phenotype, phenoAdj, hpo)
}
\arguments{
\item{phenotype}{HPO term Id <string>}

\item{phenoAdj}{A adjacency matrix (produced by the adjacency_matrix function)}

\item{hpo}{The HPO ontology data object}
}
\value{
A integer representing the relative ontology level of a term within
a connected component of a subset of the HPO.
}
\description{
When plotting subsets of the HPO data as a network plot, there is often more than
one connected component to be plotted (i.e. sections of the graph are not connected
by any edges). To map the ontology level to the size of nodes (such that high up terms
are bigger than low level terms). This was made to find the relative ontology level
with respect to other terms in its connected component so that each connected section
of the plot would have the same size for its root node. It would also be possible
to just use absolute ontology level, so that "Phenotypic abnormality" would always
be the largest datapoint etc.
}
