% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Gen.R
\name{gen_results}
\alias{gen_results}
\title{Generate rare disease EWCE results}
\usage{
gen_results(
  CTD_file = "ctd_l1l2_nz.rda",
  phenotype_to_genes_txt_file = "data/phenotype_to_genes.txt",
  output_path = "results/EWCE_output_all_phenotypes",
  reps = 1e+05,
  level = 1,
  mc.cores = 24,
  p_adj_method = "BH",
  output_merged_rda_filename = "results/Results_merged.rda",
  output_unmerged_rda_filename = "results/Results_unmerged.rda"
)
}
\arguments{
\item{CTD_file}{The CTD file}

\item{phenotype_to_genes_txt_file}{The HPO terms with gene list annotations downloaded
from the HPO website. Called phenotype_to_genes}

\item{output_path}{The directory in which results for each individual
phenotype will be saved}

\item{reps}{The number of bootstrap reps to perform with EWCE}

\item{level}{The level of cell type clustering to use in the CTD}

\item{mc.cores}{The number of CPU cores to use for paralell processing}

\item{p_adj_method}{The method to use for repeated measures in the EWCE analysis}

\item{output_merged_rda_filename}{The desired file name of the large data frame created from
all results in the output_path directory.}

\item{output_unmerged_rda_filename}{The desired file name of unmerged version of all results}
}
\value{
NULL (But saves results to desired output paths)
}
\description{
This function is called by the makefile to generate the RD EWCE results. Note,
in practice I didnt end up using the makefile and I used this function directly.
I made a shell script to load environment and run this script, as well as to
remove already completed phenotypes from the input (when having to restart the
analysis after the interactive HPC session was terminated).
I also ended up merging the individual results at the end using a separate
script, rather than this.
}
