% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_rd_ewce_results_2.R
\name{get_cell_ontology}
\alias{get_cell_ontology}
\title{Subset results and add HPO term id}
\usage{
get_cell_ontology(
  cell,
  results,
  q_threshold,
  fold_threshold,
  phenotype_to_genes,
  hpo
)
}
\arguments{
\item{cell}{The cell type of interest <string>}

\item{results}{EWCE Results <data.frame>}

\item{q_threshold}{The q value threshold of significance}

\item{fold_threshold}{The fold change threshold}

\item{phenotype_to_genes}{The HPO Ids with associated gene lists downloaded from HPO website}

\item{hpo}{The HPO Ontology data object}
}
\value{
A data frame of the selected subset of RD EWCE results with HPO ID column added.
}
\description{
This is called by the subset_phenos function to add the HPO term id column and
subset the EWCE results specific to a particular cell.
}
