% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_get_absolute_ont_level.R
\name{get_ont_level}
\alias{get_ont_level}
\title{Get absolute ontology level}
\usage{
get_ont_level(hpo, term_id)
}
\arguments{
\item{hpo}{The HPO ontology data object}

\item{term_id}{HPO term ID <string>}
}
\value{
returns the ontology level <numeric>
}
\description{
This gets the absolute ontology level of a term (without consideration for the
particular subset of the data you are looking at, as in the find_parent function).
}
