% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork_plot_gen.R
\name{ggnetwork_plot}
\alias{ggnetwork_plot}
\title{Plot RD EWCE results subset as interactive network plot}
\usage{
ggnetwork_plot(phenoNet, phenos, disease_descriptions)
}
\arguments{
\item{phenoNet}{The network object created using \code{create_network_object}}

\item{phenos}{The subset of results to be plotted (data frame)}

\item{disease_descriptions}{The data frame of all disease descriptions, This is
where new lines are added using the hpo_term_definition_list function.}
}
\value{
A interactive plot of the network of phenotypes in the selected subset of results.
}
\description{
This coerces the network object into a plot with ggnetwork (which assigns
x, y coordinates to each node in the network). The hover box with results and
disease description for each node is also added here in the
\code{pheno_ggnetwork$hover} column. Once the x and y coordiantes have been
added, it can be plot using ggplot2.
}
