% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggnetwork_plot_from_rd_ewce_results.R
\name{ggnetwork_plot_full}
\alias{ggnetwork_plot_full}
\title{Create interactive network plot start to finish}
\usage{
ggnetwork_plot_full(
  phenotype_to_genes = phenotype_to_genes,
  all_results_merged = all_results_merged,
  hpo,
  disease_descriptions,
  cell_type = "Neurons",
  q_threshold = 5e-04,
  fold_threshold = 1
)
}
\arguments{
\item{phenotype_to_genes}{The phenotype gene lists taken from the HPO}

\item{all_results_merged}{The RD EWCE Results}

\item{hpo}{The HPO ontology data object}

\item{disease_descriptions}{The dataframe of all disease descriptions in the HPO}

\item{cell_type}{The cell type of interest to be plotted}

\item{q_threshold}{The q value threshold for the subset of results to be plotted}

\item{fold_threshold}{The minimum fold change in specific expression for the subest of results to be plotted}
}
\value{
A interactive network plot of the selected subset of results from RD EWCE analysis
}
\description{
This puts all the functions together from gettig the subest of results to creating
the final interactive plot.
}
