% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hpo_get_term_definition.R
\name{hpo_get_term_definition}
\alias{hpo_get_term_definition}
\title{Get HPO term definition}
\usage{
hpo_get_term_definition(ontologyId, disease_descriptions)
}
\arguments{
\item{ontologyId}{The HPO Id of the term (string)}

\item{disease_descriptions}{A data frame of disease descriptions corresponding to each HPO Id}
}
\value{
The disease description with new lines added.
}
\description{
This gets the disease description from a data frame of disease descriptions.
The rows names of the data frame are the HPO ID and the description column is
called "description". It also adds new lines to the description so that the
hover box in the web app does not get too wide. This is done by calling the
\code{newlines_to_definition} function.
}
\examples{
\dontrun{hpo_get_term_definition("HP:123456", disease_descriptions)}

}
