% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_HPO_term_definition.R
\name{hpo_term_definition_list}
\alias{hpo_term_definition_list}
\title{Get HPO term description for a list of terms}
\usage{
hpo_term_definition_list(ontologyId_list, disease_descriptions)
}
\arguments{
\item{ontologyId_list}{A character vector of HPO Ids}

\item{disease_descriptions}{A data frame of disease descriptions for all HPO Id}
}
\description{
Need to redo this without a for loop (use \code{lapply} or something). This just
applys the \code{hpo_get_term_definition} function to a character vector of terms.
}
\examples{
\dontrun{hpo_term_definition_list(HPO_terms_char_vector, Disease_description_df)}
}
