% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix_to_network_object.R
\name{make_network_object}
\alias{make_network_object}
\title{Make network object}
\usage{
make_network_object(phenos, adjacency, hpo)
}
\arguments{
\item{phenos}{The subset of the results to be plotted}

\item{adjacency}{The adjacency matrix of all HPO terms}

\item{hpo}{The HPO ontology data object}
}
\value{
A ggnetowrk graph/ network object of a subset of the RD EWCE results.
}
\description{
This uses the network package to coerce the adjacency matrix into a
network object. It also adds the fold change, label, and relative ontology level
parameters to each node in the network.
}
\examples{
\dontrun{make_network_object(phenos,adjacency,hpo)}
}
