% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_newlines_to_hpo_term_definition.R,
%   R/hpo_add_newlines_to_term_definitions.R
\name{newlines_to_definition}
\alias{newlines_to_definition}
\title{Add new lines to disease description}
\usage{
newlines_to_definition(definition, line_length = 10)

newlines_to_definition(definition, line_length = 10)
}
\arguments{
\item{definition}{A disease description string}

\item{line_length}{A integer representing the desired words per line.}
}
\value{
The disease description with newline symbols added every nth word.

The disease description with newline symbols added every nth word.
}
\description{
Adds new lines to the description so that hover boxes dont get too wide.

Adds new lines to the description so that hover boxes dont get too wide.
}
\examples{
\dontrun{newlines_to_definition(disease_description, 10)}
\dontrun{newlines_to_definition(disease_description, 10)}
}
