% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ontoplot_network_of_rd_ewce_results_subset.R
\name{one_cell_ontology_plot_heatmap}
\alias{one_cell_ontology_plot_heatmap}
\title{Plot graph of phenotypes in the RD EWCE Results subsetted by Cell type}
\usage{
one_cell_ontology_plot_heatmap(
  results,
  cell = "Bladder cells",
  heatmapped_value = "q",
  q_threshold,
  fold_threshold,
  phenotype_to_genes,
  hpo
)
}
\arguments{
\item{results}{The RD EWCE results dataframe}

\item{cell}{The cell type of interest <string>}

\item{heatmapped_value}{"q", "fold change" or "p" <string>}

\item{q_threshold}{The q value threshold of significance}

\item{fold_threshold}{The fold change threshold}

\item{phenotype_to_genes}{The HPO Ids with associated gene lists downloaded from HPO website}

\item{hpo}{The HPO Ontology data object}
}
\value{
A ontologyPlot plot of the network of phenotypes in a subset of RD EWCE Results
}
\description{
This subsets the results, using other functions from this script, by cell type,
q value, and fold change. It then plots a graph/ network of the phenotypes and
colors the nodes by their fold change or q value (see the \code{heatmapped_value} param).
The ontologyPlot package did not have prebuilt options to create the heatmap, so the
colours are assigned manually to each phenotype in this function. There must be a more
efficent way to do this. Atleas it may be good to replace the for loop with one of the
apply functions.
}
