% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/n_enrichments_per_cell_by_branch_plot_with_hypergeometric.R
\name{plot_n_signif_phenos_per_cell_by_branch}
\alias{plot_n_signif_phenos_per_cell_by_branch}
\title{Plot number of significant enrichments per cell within specific HPO branch}
\usage{
plot_n_signif_phenos_per_cell_by_branch(
  all_results_merged,
  plot_branches = c("Abnormality of the nervous system",
    "Abnormality of the cardiovascular system", "Abnormality of the immune system"),
  q_threshold = 0.05,
  hpo = hpo,
  phenotype_to_genes = phenotype_to_genes,
  ctd = ctd
)
}
\arguments{
\item{all_results_merged}{The RD EWCE results data frame}

\item{plot_branches}{A character vector of the HPO branches to be plotted (using name not Id, maybe change this)}

\item{q_threshold}{The maximum q value accepted as a significant result.}

\item{hpo}{The HPO ontology data object (from ontologyIndex package)}

\item{phenotype_to_genes}{The .txt file containing HPO terms with corresponding gene lists}

\item{ctd}{The cell type data file for EWCE analysis.}
}
\value{
A plot of n enrichments per cell type, faceted by branch. Astrics are
used to represent significance in the hypergeometric test.
}
\description{
This plots how many significant enrichments there were per cell from within
specific branches of the HPO. It also runs hypergeometric tests to see if the
number of significant enrichments for a particular cell, within a particular branch
is higher than expected by chance. This can tell you if a branch as a whole is
commonly associated with a particular cell type.
}
\details{
This function could be split up into more smaller functions and possibly include
options for including the hypergeometric test.
}
\examples{
\dontrun{
plot_n_signif_phenos_per_cell_by_branch(all_results_merged,
plot_branches = c("Abnormality of the nervous system",
                  "Abnormality of the cardiovascular system",
                  "Abnormality of the immune system"),
                  q_threshold = 0.05,
                  hpo=hpo,
                  phenotype_to_genes=phenotype_to_genes,
                  ctd=ctd)}

}
