% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_enrichments_per_cell_type_plot.R
\name{plot_phenos_per_cell1}
\alias{plot_phenos_per_cell1}
\title{Plot number of significant enrichments per cell type in RD EWCE Results}
\usage{
plot_phenos_per_cell1(
  all_results_merged,
  cell_mappings,
  fold = 1,
  q_val = 0.005,
  dataset = "Descartes"
)
}
\arguments{
\item{all_results_merged}{RD EWCE Results <data.frame>}

\item{cell_mappings}{Mappings of cell type to tissue of origin for the CTD.}

\item{fold}{Fold change threshold for subset <numeric>}

\item{q_val}{q value threshold for subest <numeric>}

\item{dataset}{choose RNA seq "tabulamuris" or "Descartes" <string>}
}
\value{
A ggplot2 plot of number of significant results per cell type coloured by tissue
}
\description{
Plots an overview of how many phenotypes were enriched in each cell type.
}
