% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RDA_load_to_variable.R
\name{RDA_assign_load}
\alias{RDA_assign_load}
\title{Load RDA file and assign to specific variable}
\usage{
RDA_assign_load(file)
}
\arguments{
\item{file}{the path to a .rda file <string>}
}
\value{
the data contained in the .rda file
}
\description{
Can be useful, but sort of pontles now as we have decided to only use \code{.rds}
}
