% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_rd_ewce_results.R
\name{subset_phenos}
\alias{subset_phenos}
\title{Subset RD EWCE Results and add HPO Id column}
\usage{
subset_phenos(
  phenotype_to_genes,
  all_results_merged,
  hpo,
  cell_type = "Neurons",
  q_threshold = 5e-04,
  fold_threshold = 1
)
}
\arguments{
\item{phenotype_to_genes}{The list of HPO terms with their assocaited gene lists taken from HPO website}

\item{all_results_merged}{The dataframe of RD EWCE Results}

\item{hpo}{The HPO ontology data object}

\item{cell_type}{A string representing the cell type of interest.}

\item{q_threshold}{The q threshold. The subset of results will have a q lower than this}

\item{fold_threshold}{The fold change threshold. The subest of results will have a fold change greater than this.}
}
\value{
A data frame of results taken from the main data frame of results
}
\description{
This subsets  the Rare disease EWCE results by cell type, q threshold and fold change.
It calls other functions which add HPO Id column and it removes NA or invalid
HPO Ids. These checks are necessary for the interactive plots on the web app
to avoid errors, but not needed if manually plotting.
}
